analytics + bug report

+++ FlashTape/RequestPhoneViewController.m
@@ -118,7 +118,7 @@
 {
     [MBProgressHUD showHUDAddedTo:self.view animated:YES];
     [ApiManager requestSmsCode:phoneNumber retry:NO success:^(NSInteger code) {
-        NSLog(@"%lu",code);
+        NSLog(@"%lu",(long)code);
         [MBProgressHUD hideAllHUDsForView:self.view animated:YES];
         [self performSegueWithIdentifier:@"Code From Phone" sender:@[phoneNumber,[[NSNumber numberWithLong:code] stringValue]]];
     } failure:^{
